Add support for floating point fps (#866)
Related with this PR
https://github.com/ros-perception/image_pipeline/pull/723

---------

+++ image_view/include/image_view/video_recorder_node.hpp
@@ -43,7 +43,7 @@ private:
   rclcpp::Time g_last_wrote_time;
   std::string encoding;
   std::string codec;
-  int fps;
+  double fps;
   double min_depth_range;
   double max_depth_range;
   bool use_dynamic_range;
